pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rhizome"
version = "0.1.0"
description = "Децентрализованная P2P сеть на Kademlia DHT с системой seed-узлов"
readme = "ReadMe.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Rhizome Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.9.0",
    "asyncio-dgram>=2.1.2",
    "cryptography>=41.0.0",
    "pycryptodome>=3.19.0",
    "msgpack>=1.0.7",
    "protobuf>=4.25.0",
    "lmdb>=1.4.1",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "isort>=5.13.0",
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
]

[project.scripts]
rhizome = "rhizome.cli:main"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--cov=rhizome",
    "--cov-report=html",
    "--cov-report=term",
    "-v",
]

